Refactored dashboard code sturcture
